Se modifico el apartado de la creacion de los contratos, para que trajera la informacion correcta por medio del documento si existe la persona

+++ Back/WebGESCOMPAH/Entity/DTOs/Implements/Business/EstablishmentDto/EstablishmentUpdateDto.cs
@@ -1,6 +1,4 @@
 ﻿using Entity.DTOs.Base;
-using Entity.DTOs.Implements.Utilities.Images;
-    using Microsoft.AspNetCore.Http;
 
 namespace Entity.DTOs.Implements.Business.EstablishmentDto
 {
@@ -10,10 +8,8 @@ namespace Entity.DTOs.Implements.Business.EstablishmentDto
         public string Description { get; set; } = null!;
         public string Address { get; set; } = null!;
         public decimal AreaM2 { get; set; }
-        public double RentValueBase { get; set; }
+        public decimal RentValueBase { get; set; }
         public decimal UvtQty { get; set; }
-
         public int PlazaId { get; set; }
     }
-
 }
